------------------------------------------------------------------------ r194 | andrew | 2024-10-15 22:17:05 +1100 (Tue, 15 Oct 2024) | 5 lines
Found that the initial volume setting for adlib audio was initialised to zero when no config file present.

Made some small changes to fix the initial volume state and to allow for finer control of the volume level.

------------------------------------------------------------------------
r195 | andrew | 2025-01-07 12:29:22 +1100 (Tue, 07 Jan 2025) | 4 lines

I made a new file selection mechanism for the level editor that is a little
more user friendly.

------------------------------------------------------------------------