------------------------------------------------------------------------ r208 | andrew | 2025-07-28 18:42:40 +1000 (Mon, 28 Jul 2025) | 10 lines
Lots of improvements rolled up into one (due to being away)
- Made a look-up table for both multiply and divide by 10. This is to speed up conversion of screen co-ordinates to map co-ordinates and vice
  versa. The lookup tables only cover valid screen/map co-ordinates for what is on the current screen.
- added a new spider monster - it crawls around the surface of bricks including up walls and upside-down.
  It will occasionally shoot at you.
- added a new page to the help system - for hazards that are generally static. This made room for the spider and mummy on the badies screen.

I will need to test on MartyPC to see if the lookup table made any difference to how playable things are on an XT (fingers crossed)

------------------------------------------------------------------------
r209 | andrew | 2025-08-24 14:11:24 +1000 (Sun, 24 Aug 2025) | 2 lines

Added a feature to the editor (BFLEU) to change the difficulty setting. Access it by pressing escape while testing
to gain access to the testing cheat menu. I need this to make it a bit easier to test new levels are not too difficult.
------------------------------------------------------------------------