------------------------------------------------------------------------ r220 | andrew | 2025-12-01 19:40:49 +1100 (Mon, 01 Dec 2025) | 12 lines
I found there was a way to save even more memory in the llist.pas unit.

so I re-wrote much of it to push more of the data onto the heap, with smaller data sizes where I had some waste.
The memory savings are significant so it was worth it, especially if you have an XT with only 256K.

I also reworked the mechanism to select the current/next level in the episode sequence - it was a little confusing before. This is clearer.

It does however invalidate old savegames - as it will select the incorrect level when you finish the one you saved on.
If you were at the start of the episode you'll be near the end and vice versa.

I can't imagine this affects many people, but sorry if it affects you!

------------------------------------------------------------------------
r221 | andrew | 2025-12-11 18:15:42 +1100 (Thu, 11 Dec 2025) | 9 lines

I noticed a problem with the high score mechanism - essentially it doesn't know which table to store a score in if you have loaded a save game.

The reason for this is the load routine did not set the episode name in llist which is used to determine which high score table to use.

So I have re-written segments of bmenu.pas to fix this. I've changed it to use the buffer.pas file reader/writer for reading the episode list and for loading/saving the dat file for savegames.
I made a function that reads the episode list from disk and have used it to replace reading code (for the episode list) in begingame and have used it in loadgame to determine what the
episode name should be.

This should fix the issue without causing more issues with saved games.
------------------------------------------------------------------------
r222 | andrew | 2025-12-11 22:10:26 +1100 (Thu, 11 Dec 2025) | 3 lines

I re-wrote much of bconfig so that it is smaller and takes advantage of the buffer.pas buffered reader.

The code has the same functionality but is smaller.
------------------------------------------------------------------------